src/test/javascript/spec/app/entities/templates/templates-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { PlataformaTestModule } from '../../../test.module';
import { TemplatesUpdateComponent } from 'app/entities/templates/templates-update.component';
import { TemplatesService } from 'app/entities/templates/templates.service';
import { Templates } from 'app/shared/model/templates.model';

describe('Component Tests', () => {
    describe('Templates Management Update Component', () => {
        let comp: TemplatesUpdateComponent;
        let fixture: ComponentFixture<TemplatesUpdateComponent>;
        let service: TemplatesService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [PlataformaTestModule],
                declarations: [TemplatesUpdateComponent]
            })
                .overrideTemplate(TemplatesUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(TemplatesUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(TemplatesService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new Templates(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.templates = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new Templates();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.templates = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
